requirements.txt
alembic==1.7.6
anyio==3.4.0
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.0
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
ipykernel==6.6.1
ipython==7.31.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
node==0.9.28
npm==0.1.1
odict==1.8.0
optional-django==0.1.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
plumber==1.6
prompt-toolkit==3.0.24
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.1
pymongo==4.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
pyzmq==22.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.16.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3